platform_backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-dotenv>=1.0.0
cryptography>=41.0.0
httpx>=0.25.0
